Extend Swarm job/deployment to include a broker and become operational
The current Swarm setup is basically only a Swarm without a broker and
can't be used/integrated as such into CF.

This commit adds the broker support and makes the bosh release work
with a Swarm, i.e. allows to horizontally scale by adding Docker nodes.

Note of caution: Changes were done, tested, and put to production only
with our own cloud infrastructure which is neither of the supported ones.
However, the bosh-lite setup was used to develop the changes and tested.