Add logging of /proc/mounts in docker stop
We added the umount of the overlay file which we think is causing the
intermittent bug we see. However, this log line will help us debug if
that wasn't the cause and we see the problem again.

[#156017255]

Signed-off-by: Carlo Colombo <ccolombo@pivotal.io>
(cherry picked from commit ca5a324eaea97fd0a35976f9584a46c9ff387793)