Fix windows docker default config directory
This directory name appears to have been incorrectly changed with the rename to `docker-windows`. Unfortunately, the default directory appears to be hard-coded in docker and to specifically be `$env:ProgramData\docker`.

This change should allow the config to be created in docker's expected location and the config should point to bosh-managed directories.