Add logrotate property for containers
Standard docker container logs are only available for as long as
the container is running, since a monit stop also removes the
container (including logs) using docker rm.

To properly persist logs even after a container is stopped or
restarted it seems that you have to write your own application logs
and bind-mount the log directory to a persistent disk.

However, logs are not rotated this way.

To avoid the necessity of doing manual configurations to the bosh
release or on a VM, make custom logrotation settings possible via
a container property.